+++ docs/src/reference/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.3.0`
+
+- **changed** Julia requirement to LTS
+- **changed** the github actions to work on LTS and latest release
+- **added** support for `SpatialBoundaries` wombling through an extension
+- **added** `SpatialBoundaries` version 0.2 *(WEAKDEP)*
+- **added** tutorials for spatial boundaries and virtual species
+- **improved** the CI time by preventing a run of the full test suite when testing SDT
 
 ## `v1.2.4`